% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capital_raid_seasons.R
\name{coc_get_capital_raid_seasons}
\alias{coc_get_capital_raid_seasons}
\title{Get Capital Raid Seasons}
\usage{
coc_get_capital_raid_seasons(
  clan,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = get_clash_key()
)
}
\arguments{
\item{clan}{Required. Clan tag.}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}.
Default: \code{get_clash_key()}}
}
\value{
\code{tibble} with list columns for clan, clans, and history.
}
\description{
Get Capital Raid Seasons
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_capital_raid_seasons(clan = '8UC2J90Y')
\dontshow{\}) # examplesIf}
}
\concept{clan}
