% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan.R
\name{coc_get_clan}
\alias{coc_get_clan}
\title{Get a clan}
\usage{
coc_get_clan(clan, key = cr_get_key())
}
\arguments{
\item{clan}{Required. Clan tag. Default: 99R2PQVR}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}
Default: \code{get_clash_key()}}
}
\value{
a \code{tibble} with each row as a player
}
\description{
Gets full clan details
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_clan('99R2PQVR')
\dontshow{\}) # examplesIf}
}
\concept{clan}
