% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan_search.R
\name{coc_get_clan_search}
\alias{coc_get_clan_search}
\title{Search Clans}
\usage{
coc_get_clan_search(
  name = NULL,
  location_id = NULL,
  war_frequency = NULL,
  min_members = NULL,
  max_members = NULL,
  min_clan_points = NULL,
  min_clan_level = NULL,
  limit = NULL,
  after = NULL,
  before = NULL,
  label_ids = NULL,
  key = get_clash_key()
)
}
\arguments{
\item{name}{Search clans by name}

\item{location_id}{Filter by clan location identifier}

\item{min_members}{Filter by minimum number of clan members}

\item{max_members}{Filter by maximum number of clan members}

\item{min_clan_points}{minimum amount of clan points}

\item{min_clan_level}{minimum clan level}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{label_ids}{comma separated vector to filter results by}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}.
Default: \code{get_clash_key()}}
}
\value{
a \code{tibble} with each row as a player
}
\description{
Search Clans
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_clan_search('Dusty Death', limit = 5)
\dontshow{\}) # examplesIf}
}
\concept{clan}
