% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_goldpass.R
\name{coc_get_goldpass}
\alias{coc_get_goldpass}
\title{Get Gold Pass Season Information}
\usage{
coc_get_goldpass(key = get_clash_key())
}
\arguments{
\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
a \code{tibble} of gold pass info
}
\description{
Get Gold Pass Season Information
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_goldpass()
\dontshow{\}) # examplesIf}
}
\concept{goldpass}
