% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league.R
\name{coc_get_league}
\alias{coc_get_league}
\title{Get information for one league}
\usage{
coc_get_league(
  league,
  season,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = coc_get_key()
)
}
\arguments{
\item{league}{league id, such as 29000022}

\item{season}{season id, such as '2022-09'}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}.
Default: \code{get_clash_key()}}
}
\value{
\code{tibble} with list columns for clan, clans, and history.
}
\description{
Get information for one league
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_league(league = '29000022', season = '2022-09', limit = 10)
\dontshow{\}) # examplesIf}
}
\concept{clan}
