% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{coc_get_locations}
\alias{coc_get_locations}
\title{Get Locations}
\usage{
coc_get_locations(
  limit = NULL,
  after = NULL,
  before = NULL,
  key = coc_get_key()
)
}
\arguments{
\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}.
Default: \code{get_clash_key()}}
}
\value{
tibble of card info
}
\description{
Get Locations
}
\examples{
\dontshow{if (clash::coc_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_locations()
\dontshow{\}) # examplesIf}
}
\concept{locations}
