% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player.R
\name{coc_get_player}
\alias{coc_get_player}
\title{Get a single player}
\usage{
coc_get_player(tag = "92GCQQYP", key = get_clash_key())
}
\arguments{
\item{tag}{Required. The player to return.}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}
Default: \code{coc_get_key()}}
}
\value{
a \code{tibble} of player info
}
\description{
Gets the player details for one player.
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_player('92GCQQYP')
\dontshow{\}) # examplesIf}
}
\concept{player}
