% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_labels.R
\name{coc_get_player_labels}
\alias{coc_get_player_labels}
\title{Get player labels}
\usage{
coc_get_player_labels(
  limit = NULL,
  after = NULL,
  before = NULL,
  key = get_clash_key()
)
}
\arguments{
\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
tibble of card info
}
\description{
Get player labels
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_player_labels()
\dontshow{\}) # examplesIf}
}
\concept{labels}
