% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_war_leagues.R
\name{coc_get_war_leagues}
\alias{coc_get_war_leagues}
\title{Get War Leagues}
\usage{
coc_get_war_leagues(
  limit = NULL,
  after = NULL,
  before = NULL,
  key = coc_get_key()
)
}
\arguments{
\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashofclans.com/#/documentation}.
Default: \code{get_clash_key()}}
}
\value{
\code{tibble} with list columns for clan, clans, and history.
}
\description{
Get War Leagues
}
\examples{
\dontshow{if (clash::has_clash_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
coc_get_war_leagues(limit = 10)
\dontshow{\}) # examplesIf}
}
\concept{clan}
