% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\alias{has_clash_key}
\alias{get_clash_key}
\alias{coc_get_key}
\alias{coc_set_key}
\alias{coc_has_key}
\title{Check or Get Clash of Clans API Key}
\usage{
has_clash_key()

get_clash_key()

coc_get_key()

coc_set_key(key, overwrite = FALSE, install = FALSE)

coc_has_key()
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Check or Get Clash of Clans API Key
}
\examples{
has_clash_key()
}
